Fix: Repeat 'evil-scroll-line-up/down' in *Cape*'s insert state
Add evil-search module to Cape

Bind 'evil-scroll-line-up/down' to 'C-e/C-y' for Cape's insert state.

Previously, it was possible to bind those keys to 'evil-scroll-line-up/down' globally, but not while in Cape's insert state. This was because Cape's insert state had its own keymap (cape-global-map) which did not include those bindings.

This commit fixes that issue by adding a repeat map to Cape's insert state which includes bindings for 'evil-scroll-line-up/down'. This allows those keys to be repeated in Cape's insert state, even if they are not bound globally.

Add `evil-search` module to Cape

This adds the `evil-search` module to Cape, which provides support for using Evil's search commands in Cape. This allows users to take advantage of Evil's powerful search functionality within Cape, such as incremental search, motion commands, and search operators.